server/requirements.txt
flask
flask-cors
google-generativeai
pinecone
python-dotenv
requests
cryptography
# gunicorn is for unix production, use waitress for windows or flask dev server
# streamlit is removed as we use react frontend
